pyproject.toml
[project]
name = "miiment-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "boto3>=1.37.36",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic[email]>=2.11.3",
    "pymongo>=4.12.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sentry-sdk>=2.26.1",
    "uvicorn>=0.34.1",
]
